+++ src/plugin.svelte
@@ -1,10 +1,8 @@
-<svelte:options tag={null} />
-
 <script lang="ts">
     import PluginStore from "./plugin-store.svelte";
     import { name, version } from "../package.json";
 
-    let isOpen = false;
+    let isOpen = $state(false);
 
     export async function run() {
         isOpen = true;