+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2025-10-29
+
+### Added
+- Added Documentation how to test in the OpenSCD instance
+
+### Changed
+- Migrated to Svelte 5
+- `installed` is not a property of Plugin anymore
+
 ## [1.0.0] - 2024-11-18
 
 ### Added